fix UniformSampleFrames and UT
