+++ mmaction/datasets/transforms/pose_loading.py
@@ -15,141 +15,6 @@ from .formatting import Rename
 
 
 @TRANSFORMS.register_module()
-class UniformSampleFrames(BaseTransform):
-    """Uniformly sample frames from the video.
-
-    To sample an n-frame clip from the video. UniformSampleFrames basically
-    divide the video into n segments of equal length and randomly sample one
-    frame from each segment. To make the testing results reproducible, a
-    random seed is set during testing, to make the sampling results
-    deterministic.
-
-    Required keys are ``'total_frames'``, ``'start_index'`` , added or
-    modified keys are ``'frame_inds'``, ``'clip_len'``,
-    ``'frame_interval'`` and ``'num_clips'``.
-
-    Args:
-        clip_len (int): Frames of each sampled output clip.
-        num_clips (int): Number of clips to be sampled. Defaults to 1.
-        test_mode (bool): Store True when building test or validation dataset.
-            Defaults to False.
-        seed (int): The random seed used during test time. Defaults to 255.
-    """
-
-    def __init__(self, clip_len, num_clips=1, test_mode=False, seed=255):
-
-        self.clip_len = clip_len
-        self.num_clips = num_clips
-        self.test_mode = test_mode
-        self.seed = seed
-
-    def _get_train_clips(self, num_frames, clip_len):
-        """Uniformly sample indices for training clips.
-
-        Args:
-            num_frames (int): The number of frames.
-            clip_len (int): The length of the clip.
-        """
-
-        assert self.num_clips == 1
-        if num_frames < clip_len:
-            start = np.random.randint(0, num_frames)
-            inds = np.arange(start, start + clip_len)
-        elif clip_len <= num_frames < 2 * clip_len:
-            basic = np.arange(clip_len)
-            inds = np.random.choice(
-                clip_len + 1, num_frames - clip_len, replace=False)
-            offset = np.zeros(clip_len + 1, dtype=np.int32)
-            offset[inds] = 1
-            offset = np.cumsum(offset)
-            inds = basic + offset[:-1]
-        else:
-            bids = np.array(
-                [i * num_frames // clip_len for i in range(clip_len + 1)])
-            bsize = np.diff(bids)
-            bst = bids[:clip_len]
-            offset = np.random.randint(bsize)
-            inds = bst + offset
-        return inds
-
-    def _get_test_clips(self, num_frames, clip_len):
-        """Uniformly sample indices for testing clips.
-
-        Args:
-            num_frames (int): The number of frames.
-            clip_len (int): The length of the clip.
-        """
-
-        np.random.seed(self.seed)
-        if num_frames < clip_len:
-            # Then we use a simple strategy
-            if num_frames < self.num_clips:
-                start_inds = list(range(self.num_clips))
-            else:
-                start_inds = [
-                    i * num_frames // self.num_clips
-                    for i in range(self.num_clips)
-                ]
-            inds = np.concatenate(
-                [np.arange(i, i + clip_len) for i in start_inds])
-        elif clip_len <= num_frames < clip_len * 2:
-            all_inds = []
-            for i in range(self.num_clips):
-                basic = np.arange(clip_len)
-                inds = np.random.choice(
-                    clip_len + 1, num_frames - clip_len, replace=False)
-                offset = np.zeros(clip_len + 1, dtype=np.int32)
-                offset[inds] = 1
-                offset = np.cumsum(offset)
-                inds = basic + offset[:-1]
-                all_inds.append(inds)
-            inds = np.concatenate(all_inds)
-        else:
-            bids = np.array(
-                [i * num_frames // clip_len for i in range(clip_len + 1)])
-            bsize = np.diff(bids)
-            bst = bids[:clip_len]
-            all_inds = []
-            for i in range(self.num_clips):
-                offset = np.random.randint(bsize)
-                all_inds.append(bst + offset)
-            inds = np.concatenate(all_inds)
-        return inds
-
-    def transform(self, results):
-        """Perform the SampleFrames loading.
-
-        Args:
-            results (dict): The resulting dict to be modified and passed
-                to the next transform in pipeline.
-        """
-        num_frames = results['total_frames']
-
-        if self.test_mode:
-            inds = self._get_test_clips(num_frames, self.clip_len)
-        else:
-            inds = self._get_train_clips(num_frames, self.clip_len)
-
-        inds = np.mod(inds, num_frames)
-        start_index = results['start_index']
-        inds = inds + start_index
-
-        results['frame_inds'] = inds.astype(np.int32)
-        results['clip_len'] = self.clip_len
-        results['frame_interval'] = None
-        results['num_clips'] = self.num_clips
-        return results
-
-    def __repr__(self):
-        repr_str = (f'{self.__class__.__name__}('
-                    f'clip_len={self.clip_len}, '
-                    f'num_clips={self.num_clips}, '
-                    f'test_mode={self.test_mode}, '
-                    f'seed={self.seed})')
-        return repr_str
-
-
-@TRANSFORMS.register_module()
 class PoseDecode(BaseTransform):
     """Load and decode pose with given indices.
